cpu-align: Refactor code to avoid using hard-coded values
The focus of this patch is test cases that validate the
behavior of the scheduer when the values of a single resource
is the only difference between two nodes. This differnce between
the candidate nodes and the target node ultimately influences
the node selection.

On candidate nodes that should be disregarded by scheduler, we
simulate scenarios where a specific resource is insufficient to
meet the resource requirement of the pod.

This patch performs code refactoring to evaluate free resources
that should be left on candidate nodes (non-targetted nodes)
rather than using hardcoded values.

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>