test-deletion: Add Manifests that were used to test deletion scenario
These manifests are very specific to the cluster where they were tested
as they correspond to padder pods that are supposed to be deployed on specific
nodes. We therefore, use nodeselector to indicate the nodes where the pods should
be deployed.

Typically, the order in which they are deployed matter as Topology Manager and
CPU Manager allocate resource sequentially. So, deploy the pods in the following
sequence:

example-deployment-node0-NUMA0-padder.yaml
example-deployment-node0-NUMA1-padder.yaml
example-deployment-node1-NUMA0-padder.yaml
example-deployment-node1-NUMA0-padder.yaml

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>